Fixed a compiler warning.

+++ rtos/fs/fs_buffer.c
@@ -866,7 +866,7 @@ int32_t fs_buffer_pull_offset(FS_BUFFER *buffer, void *data, uint32_t size, uint
     int32_t status = SUCCESS;
     uint32_t this_size, this_offset = offset;
 #ifdef OS_LITTLE_ENDIAN
-    uint8_t reverse = (((flags & FS_BUFFER_PACKED) != 0) ^ ((flags & FS_BUFFER_HEAD) != 0));
+    uint8_t reverse = (uint8_t)(((flags & FS_BUFFER_PACKED) != 0) ^ ((flags & FS_BUFFER_HEAD) != 0));
 #endif
     uint8_t *to;
 
@@ -1047,7 +1047,7 @@ int32_t fs_buffer_push_offset(FS_BUFFER *buffer, void *data, uint32_t size, uint
     FS_BUFFER_ONE *one = NULL;
     uint32_t this_size, this_offset = offset;
 #ifdef OS_LITTLE_ENDIAN
-    uint8_t reverse = (((flags & FS_BUFFER_PACKED) != 0) ^ ((flags & FS_BUFFER_HEAD) != 0));
+    uint8_t reverse = (uint8_t)(((flags & FS_BUFFER_PACKED) != 0) ^ ((flags & FS_BUFFER_HEAD) != 0));
 #endif
     uint8_t *from;
